ENH: Fold any instruction spanning on several lines
This concerns instructions as well as function declarations.

This commit permits to merge comments with single-line function declarations,
but also to fold multi-line instructions. At this point, I don't see how to
distinguish the two cases: functions calls and function declarations are quite
alike.

Should it be an option?

Relates to #6